Move networksetupconfig to the config file
